// frontend/lib/admin-types.ts
export interface Certification {
    id?: number;
    name: string;
    issuer: string;
    date: string;
    credential?: string;
    validity?: string;
    badge?: string;
    verify_url?: string;
}